created base and util files
